Correct rounding in getNumShards

+++ src/java/org/apache/cassandra/db/compaction/unified/StaticController.java
@@ -54,7 +54,7 @@ public class StaticController extends Controller
                             boolean ignoreOverlapsInExpirationCheck,
                             boolean l0ShardsEnabled,
                             int baseShardCount,
-                            double targetSStableSize,
+                            long targetSStableSize,
                             double sstableGrowthModifier,
                             int reservedThreadsPerLevel,
                             OverlapInclusionMethod overlapInclusionMethod)
@@ -89,7 +89,7 @@ public class StaticController extends Controller
                                   boolean ignoreOverlapsInExpirationCheck,
                                   boolean l0ShardsEnabled,
                                   int baseShardCount,
-                                  double targetSStableSize,
+                                  long targetSStableSize,
                                   double sstableGrowthModifier,
                                   int reservedThreadsPerLevel,
                                   OverlapInclusionMethod overlapInclusionMethod,
